Cleanup, fixed redux state
