--- src/components/NavItems/tools/Metadata/Hooks/useVideoTreatment.jsx
@@ -1,74 +1,0 @@
-import { useEffect } from "react";
-import { useDispatch } from "react-redux";
-
-import * as mp4box from "mp4box";
-import { setError } from "redux/reducers/errorReducer";
-
-import {
-  setMetadataLoading,
-  setMetadataResult,
-} from "../../../../../redux/reducers/tools/metadataReducer";
-
-const useVideoTreatment = (mediaUrl, keyword) => {
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    const handleError = (error) => {
-      if (keyword(error) !== "") dispatch(setError(keyword(error)));
-      else dispatch(setError("Unknown error"));
-      dispatch(setMetadataLoading(false));
-    };
-
-    let videoTreatment = () => {
-      let video = mp4box.createFile();
-
-      video.onReady = (info) => {
-        dispatch(
-          setMetadataResult({
-            url: mediaUrl,
-            result: info,
-            notification: false,
-            loading: false,
-            isImage: false,
-          }),
-        );
-      };
-
-      video.onError = (error) => {
-        console.error("mp4 error : " + error);
-        handleError("metadata_table_error");
-      };
-
-      let fileReader = new FileReader();
-      fileReader.onload = () => {
-        let arrayBuffer = fileReader.result;
-
-        arrayBuffer.fileStart = 0;
-        video.appendBuffer(arrayBuffer);
-        video.flush();
-      };
-      fileReader.onerror = () => {
-        handleError("metadata_table_error");
-      };
-
-      let blob = null;
-      let xhr = new XMLHttpRequest();
-
-      xhr.open("GET", mediaUrl);
-      xhr.responseType = "blob";
-      xhr.onload = function () {
-        blob = xhr.response; //xhr.response is now a blob object
-        fileReader.readAsArrayBuffer(blob);
-      };
-      xhr.send();
-      xhr.onreadystatechange = () => {
-        if (xhr.readyState === 4 && xhr.status !== 200) {
-          handleError("metadata_table_error");
-        }
-      };
-    };
-
-    if (mediaUrl) videoTreatment();
-  }, [mediaUrl, keyword, dispatch]);
-};
-export default useVideoTreatment;
