+++ src/components/NavItems/tools/Metadata/Metadata.jsx
@@ -1,7 +1,6 @@
-import { useMutation } from "@tanstack/react-query";
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { useLocation, useParams } from "react-router-dom";
+import { useParams } from "react-router-dom";
 
 import Alert from "@mui/material/Alert";
 import Box from "@mui/material/Box";
@@ -11,15 +10,12 @@ import Fade from "@mui/material/Fade";
 import Stack from "@mui/material/Stack";
 
 import { getclientId } from "@Shared/GoogleAnalytics/MatomoAnalytics";
-import { isValidUrl } from "@Shared/Utils/URLUtils";
 import {
   getFileTypeFromFileObject,
   getFileTypeFromUrl,
 } from "@Shared/Utils/fileUtils";
-import { parseMetadata } from "@uswriting/exiftool";
 import { i18nLoadNamespace } from "components/Shared/Languages/i18nLoadNamespace";
 import StringFileUploadField from "components/Shared/StringFileUploadField";
-import exifr from "exifr";
 
 import { useTrackEvent } from "../../../../Hooks/useAnalytics";
 import { imageMetadata as imageMetadataTool } from "../../../../constants/tools";
@@ -32,10 +28,14 @@ import HeaderTool from "../../../Shared/HeaderTool/HeaderTool";
 import { KNOWN_LINKS } from "../../Assistant/AssistantRuleBook";
 import MetadataImageResult from "./Results/MetadataImageResult";
 import MetadataVideoResult from "./Results/MetadataVideoResult";
+import {
+  getImageMetadataFromFile,
+  getImageMetadataFromUrl,
+} from "./api/imageMetadataApi";
+import { useVideoMetadataMutation } from "./hooks/useVideoMetadataMutation";
 
 const Metadata = () => {
   const { url, type } = useParams();
-  const location = useLocation();
 
   const keyword = i18nLoadNamespace("components/NavItems/tools/Metadata");
 
@@ -53,7 +53,7 @@ const Metadata = () => {
 
   const [input, setInput] = useState(resultUrl ? resultUrl : "");
   const [fileInput, setFileInput] = useState(null);
-  const [imageUrl, setImageurl] = useState(null);
+  const [imageUrl, setImageUrl] = useState(null);
   const [videoUrl, setVideoUrl] = useState(null);
   const [urlDetected, setUrlDetected] = useState(false);
   const [error, setError] = useState(false);
@@ -61,58 +61,6 @@ const Metadata = () => {
   const [imageMetadata, setImageMetadata] = useState(
     resultData ? resultData : null,
   );
-
-  const exifrOptions = {
-    exif: true,
-    gps: true,
-    iptc: true,
-    jfif: true,
-    tiff: true,
-    mergeOutput: false,
-  };
-
-  const extractVideoMetadata = async (url) => {
-    if (!isValidUrl(url)) {
-      throw new Error("Invalid URL");
-    }
-
-    // Fetch the video
-    const response = await fetch(url, {
-      mode: "cors",
-    });
-
-    if (!response.ok) {
-      throw new Error(`HTTP error ${response.status}`);
-    }
-
-    const blob = await response.blob();
-
-    // Parse the metadata
-    const jsonMetadata = await parseMetadata(
-      new File([blob], "video-file", {
-        type: blob.type,
-      }),
-      {
-        args: ["-a", "-g1", "-json", "-n"],
-        transform: (data) => JSON.parse(data),
-      },
-    );
-
-    if (!jsonMetadata.success) {
-      throw new Error("No metadata found.");
-    }
-
-    return jsonMetadata.data[0];
-  };
-
-  const getVideoMetadata = useMutation({
-    mutationFn: (url) => {
-      return extractVideoMetadata(url);
-    },
-    onSuccess: (data) => {
-      setMetadataResult(data);
-    },
-  });
 
   const client_id = getclientId();
   useTrackEvent(
@@ -134,53 +82,30 @@ const Metadata = () => {
     uid,
   );
 
-  async function getImageMetadataFromUrl(url) {
-    try {
-      // Validate the URL format
-      if (!url || typeof url !== "string" || !/^https?:\/\//.test(url)) {
-        return new Error("Invalid URL provided");
-      }
+  const getVideoMetadata = useVideoMetadataMutation({
+    onSuccess: (data) => {
+      dispatch(
+        setMetadataResult({
+          url: videoUrl,
+          result: data instanceof Error ? null : data,
+          notification: false,
+          loading: false,
+          isImage: false,
+        }),
+      );
+    },
+  });
 
-      // Fetch the image
-      const response = await fetch(url);
-
-      // Check for a successful response
-      if (!response.ok) {
-        return new Error(
-          `Failed to fetch image: ${response.status} ${response.statusText}`,
-        );
-      }
-
-      // Check if the content-type is an image
-      const contentType = response.headers.get("Content-Type");
-      if (!contentType || !contentType.startsWith("image/")) {
-        return new Error("The provided URL is not an image");
-      }
-
-      // Convert response to a Blob
-      const blob = await response.blob();
-
-      // Extract metadata using exifr
-      const metadata = await exifr.parse(blob, exifrOptions);
-
-      // Handle missing metadata
-      if (!metadata) {
-        return new Error("No EXIF metadata found in the image");
-      }
-
-      return metadata;
-    } catch (error) {
-      console.error("Error extracting metadata:", error.message);
-      return null; // Return null instead of throwing to prevent app crashes
-    }
-  }
-
-  const submitUrl = async () => {
-    // Reset state
-    cleanMetadataState();
+  const resetMetadataState = () => {
+    dispatch(cleanMetadataState());
     getVideoMetadata.reset();
     setImageMetadata(null);
     setError(false);
+    setImageUrl(null);
+  };
+
+  const submitUrl = async () => {
+    resetMetadataState();
 
     try {
       if (!input && !fileInput) {
@@ -207,12 +132,12 @@ const Metadata = () => {
       if (fileType.mime.includes("image")) {
         // Set the image URL
         const imageUrl = input || URL.createObjectURL(fileInput);
-        setImageurl(imageUrl);
+        setImageUrl(imageUrl);
 
         // Extract metadata
         const metadata = input
           ? await getImageMetadataFromUrl(input)
-          : await exifr.parse(fileInput, exifrOptions);
+          : await getImageMetadataFromFile(fileInput);
 
         setImageMetadata(metadata instanceof Error ? null : metadata);
 
@@ -229,11 +154,6 @@ const Metadata = () => {
         return;
       }
 
-      // if (fileType.mime.includes("video")) {
-      //   setVideoUrl(input || URL.createObjectURL(fileInput));
-      //   return;
-      // }
-
       throw new Error("Unsupported file type");
     } catch (error) {
       console.error("Error in submitUrl:", error.message);
@@ -242,7 +162,7 @@ const Metadata = () => {
   };
 
   useEffect(() => {
-    setImageurl(null);
+    setImageUrl(null);
   }, [imageUrl]);
 
   useEffect(() => {
@@ -252,22 +172,7 @@ const Metadata = () => {
     }
   }, [urlDetected]);
 
-  const [initTool, setInitTool] = useState(true);
-
   const dispatch = useDispatch();
-
-  if (initTool) {
-    if (location.state != null) {
-      if (location.state.media === "image") {
-        dispatch(setMetadataMediaType("image"));
-      } else if (location.state.media === "video") {
-        dispatch(setMetadataMediaType("video"));
-      }
-    } else {
-      // console.log(mediaType);
-    }
-    setInitTool(false);
-  }
 
   useEffect(() => {
     if (url && url !== KNOWN_LINKS.OWN) {
@@ -293,6 +198,7 @@ const Metadata = () => {
 
   const handleCloseFile = () => {
     setFileInput(null);
+    resetMetadataState();
   };
 
   return (
@@ -351,29 +257,18 @@ const Metadata = () => {
         <Alert severity="error">{keyword("metadata_generic_error")}</Alert>
       )}
 
-      {getVideoMetadata.isSuccess && videoUrl && (
-        <MetadataVideoResult
-          metadata={getVideoMetadata.data}
-          videoSrc={videoUrl}
-          handleClose={handleCloseResult}
-        />
+      {resultData && !resultIsImage && resultUrl && (
+        <MetadataVideoResult metadata={resultData} videoSrc={resultUrl} />
       )}
 
-      {resultData ? (
-        resultIsImage ? (
-          <MetadataImageResult
-            result={resultData}
-            metadata={imageMetadata}
-            closeResult={handleCloseResult}
-            imageSrc={resultUrl}
-          />
-        ) : (
-          <MetadataVideoResult
-            result={resultData}
-            closeResult={handleCloseResult}
-          />
-        )
-      ) : null}
+      {resultData && resultIsImage && (
+        <MetadataImageResult
+          result={resultData}
+          metadata={imageMetadata}
+          closeResult={handleCloseResult}
+          imageSrc={resultUrl}
+        />
+      )}
     </Box>
   );
 };
