+++ src/components/NavItems/tools/Metadata/hooks/useVideoMetadataMutation.jsx
@@ -1,0 +1,10 @@
+import { useMutation } from "@tanstack/react-query";
+
+import { extractVideoMetadata } from "../api/videoMetadataApi";
+
+export const useVideoMetadataMutation = (options = {}) => {
+  return useMutation({
+    mutationFn: extractVideoMetadata,
+    ...options,
+  });
+};
